requirements.txt
click
whisperx
Pillow
ffmpeg-python
pyyaml
pandas
google-genai
ftfy
regex
toolz
matplotlib
